refactor: GNB NavBar로 이름 변경
