--- components/organisms/GNB/index.ts
@@ -1,1 +1,0 @@
-export { GNB } from './GNB';
