+++ components/organisms/NavBar/NavBar.stories.tsx
@@ -1,0 +1,11 @@
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { NavBar } from './NavBar';
+
+export default {
+  title: 'Components/Molecules/NavBar',
+  component: NavBar,
+} as ComponentMeta<typeof NavBar>;
+
+export const Default: ComponentStory<typeof NavBar> = () => {
+  return <NavBar />;
+};
